+++ Quake/cd_win.c
@@ -401,7 +401,7 @@ static void CD_f (void)
 
 LONG CDAudio_MessageHandler(HWND hWnd, UINT uMsg, WPARAM wParam, LPARAM lParam)
 {
-	if (lParam != wDeviceID)
+	if (lParam != (LONG)wDeviceID)
 		return 1;
 
 	switch (wParam)
